requirements.txt
altgraph==0.17.2
certifi==2022.5.18.1
future==0.18.2
pefile==2021.9.3
Pillow==9.1.1
pip==21.2.4
pyinstaller==5.1
pyinstaller-hooks-contrib==2022.5
pywin32-ctypes==0.2.0
setuptools==61.2.0
tkScrolledFrame==1.0.4
wheel==0.37.1
wincertstore==0.2
